Refactoring TCT-ColBERT regressions.
